+++ session_test.go
@@ -77,7 +77,7 @@ func TestNilNegotiatorPanics(t *testing.T) {
 			t.Error("Expected panic, did not get one")
 		}
 	}()
-	xmpp.NegotiateSession(context.Background(), jid.JID{}, jid.JID{}, nil, false, nil)
+	xmpp.NegotiateSession(context.Background(), jid.JID{}, jid.JID{}, nil, nil)
 }
 
 var errTestNegotiate = errors.New("a test error")
@@ -146,7 +146,7 @@ func TestNegotiator(t *testing.T) {
 				Reader: strings.NewReader(tc.in),
 				Writer: buf,
 			}
-			session, err := xmpp.NegotiateSession(context.Background(), tc.location, tc.origin, rw, false, tc.negotiator)
+			session, err := xmpp.NegotiateSession(context.Background(), tc.location, tc.origin, rw, tc.negotiator)
 			if ((err == nil || tc.err == nil) && (err != nil || tc.err != nil)) && err.Error() != tc.err.Error() {
 				t.Errorf("unexpected error: want=%q, got=%q", tc.err, err)
 			}
@@ -399,11 +399,11 @@ func TestNegotiateStreamError(t *testing.T) {
 	clientConn, serverConn := net.Pipe()
 	clientJID := jid.MustParse("me@example.net")
 	go func() {
-		xmpp.NegotiateSession(ctx, clientJID.Bare(), clientJID, serverConn, true, xmpp.NewNegotiator(xmpp.StreamConfig{
+		xmpp.ReceiveSession(ctx, serverConn, xmpp.NewNegotiator(xmpp.StreamConfig{
 			Features: []xmpp.StreamFeature{errorStartTLS(stream.Conflict)},
 		}))
 	}()
-	_, err := xmpp.NegotiateSession(ctx, clientJID, clientJID.Bare(), clientConn, false, xmpp.NewNegotiator(xmpp.StreamConfig{
+	_, err := xmpp.NegotiateSession(ctx, clientJID, clientJID.Bare(), clientConn, xmpp.NewNegotiator(xmpp.StreamConfig{
 		Features: []xmpp.StreamFeature{xmpp.StartTLS(nil)},
 	}))
 	if !errors.Is(err, stream.Conflict) {