+++ websocket/integration_test.go
@@ -71,7 +71,7 @@ func integrationDialWebsocket(ctx context.Context, t *testing.T, cmd *integratio
 		t.Fatalf("error dialing WebSocket connection: %v", err)
 	}
 	session, err := xmpp.NegotiateSession(
-		context.TODO(), j.Domain(), j, conn, false,
+		context.TODO(), j.Domain(), j, conn,
 		xmpp.NewNegotiator(xmpp.StreamConfig{
 			WebSocket: true,
 			Secure:    true,